backend/requirements.txt
fastapi==0.110.1
uvicorn==0.29.0
websockets==12.0
pydantic==2.6.4
python-dotenv==1.0.1
openai==1.14.3
faiss-cpu
numpy
